More work
Implemented the audio chopping section. It'll slice up a wav file
according to the timestamps in a .cha file. Also reworked the time
stuff so that we can have millisecond accuracy when making the slices.
All that's left is the clips playback (which is basically already
implemented) and the clip classification stuff.